requirements.txt
numpy==1.23.3
pandas==1.5.0
psycopg2==2.9.3
pyarrow==9.0.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
grpcio-tools ~= 1.50
django-environ ~= 0.9
redis ~= 4.3
